Remove rabbitmq dependency
There are two issues with this dependency.

(1) This dependency is wrong and can't be deployed by
librarian-puppet[1].

(2) We are using our custom rabbitmq module and this won't work
for us even if (1) is fixed.

Remove this dependency in stein, and plan for moving to upstream
puppet/rabbitmq module in train.

[1]: https://bugs.launchpad.net/puppet-nova/+bug/1849768

Change-Id: I0a0e13a75b7d5a7689166686e0360c875b086583